fix(fe): auto refresh access token only in client side
